+++ lib/vagrant-multiprovider-snap/plugin.rb
@@ -9,22 +9,41 @@ module VagrantSnap
         require_relative "providers/virtualbox/driver/base"
         require_relative "providers/virtualbox/driver/meta"
 
+	require_relative "providers/hyperv/action"
+	require_relative "providers/hyperv/driver/base"
+
         begin
 
-            # Make sure the fusion plugin is installed, then
-            # include the fusion providers.
+            # Make sure the fusion plugin is installed (explicitly
+            # requiring the classes we're going to extend otherwise
+            # they don't get autloaded in time for us to do so)
             
-            Vagrant.require_plugin("vagrant-vmware-fusion")
+            require "vagrant-vmware-fusion/action" 
+            require "vagrant-vmware-fusion/driver" 
+
+            # ...then include our overrides
+            
             require_relative "providers/vmware_fusion/action"
             require_relative "providers/vmware_fusion/driver/base"
 
-        rescue Vagrant::Errors::PluginLoadError => e
-
+        rescue LoadError => e
             # If we can't load the fusion plugin, quietly ignore it
-            # I'd like to log an info level message here, but I don't
-            # think I can get at a ui object in this scope.
-
         end
+
+        begin
+
+            # Same again but for the workstation plugin
+
+            require "vagrant-vmware-workstation/action" 
+            require "vagrant-vmware-workstation/driver" 
+
+            require_relative "providers/vmware_workstation/action"
+            require_relative "providers/vmware_workstation/driver/base"
+
+        rescue LoadError => e
+            # If we can't load the workstation plugin, quietly ignore it
+        end
+
 
         name "snap command"