+++ lib/vagrant-multiprovider-snap/providers/vmware_workstation/action/has_snapshot.rb
@@ -1,6 +1,6 @@
 module HashiCorp
 
-    module VagrantVMwarefusion
+    module VagrantVMwareworkstation
 
         module Action
 
@@ -13,12 +13,10 @@ module HashiCorp
                 def call(env)
 
                     if env[:snap_name].nil?
-                        snap_name = env[:machine].provider.driver.snapshot_list.last
-                    else
-                        snap_name = env[:snap_name]
+                        env[:snap_name] = env[:machine].provider.driver.snapshot_list.last
                     end
 
-                    env[:result] = env[:machine].provider.driver.has_snapshot?(snap_name)
+                    env[:result] = env[:machine].provider.driver.has_snapshot?(env[:snap_name])
 
                     @app.call(env)