+++ lib/vagrant-multiprovider-snap/providers/vmware_workstation/action/snapshot_rollback.rb
@@ -1,6 +1,6 @@
 module HashiCorp
 
-    module VagrantVMwarefusion
+    module VagrantVMwareworkstation
 
         module Action